lib/material/test/requirements.txt
robotframework==5.0.1
pyserial==3.4
